refactor: separate chain_state and proposal_table from shared
