+++ chain/src/chain.rs
@@ -6,9 +6,10 @@ use ckb_core::transaction::{OutPoint, ProposalShortId};
 use ckb_core::BlockNumber;
 use ckb_db::batch::Batch;
 use ckb_notify::NotifyController;
+use ckb_shared::chain_state::ChainState;
 use ckb_shared::error::SharedError;
 use ckb_shared::index::ChainIndex;
-use ckb_shared::shared::{ChainState, Shared};
+use ckb_shared::shared::Shared;
 use ckb_shared::txo_set::TxoSetDiff;
 use ckb_traits::ChainProvider;
 use ckb_verification::{BlockVerifier, TransactionsVerifier, Verifier};
